Implement behavior step definitions
