+++ tests/behavior/steps/backup_roundtrip_steps.py
@@ -1,8 +1,6 @@
-"""Placeholder steps for backup roundtrip feature."""
+"""Backup roundtrip scenario definitions."""
 
-import pytest
 from pytest_bdd import scenarios
 
-pytestmark = pytest.mark.pending(reason="Step implementations pending")
 
 scenarios("../features/backup_roundtrip.feature")
